+++ geo-types/src/line_string.rs
@@ -201,6 +201,7 @@ impl<T: CoordinateType> LineString<T> {
     /// let line_string: LineString<f32> = coords.into_iter().collect();
     /// assert_eq!(3, line_string.num_coords());
     /// ```
+    #[deprecated(note = "Use geo::algorithm::coords_iter::CoordsIter::coords_count instead")]
     pub fn num_coords(&self) -> usize {
         self.0.len()
     }