+++ geo/src/algorithm/coords_iter.rs
@@ -123,8 +123,8 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for Polygon<T> {
 
     /// Return the number of coordinates in the `Polygon`.
     fn coords_count(&'a self) -> usize {
-        self.exterior().num_coords() +
-            self.interiors().iter().map(|i| i.num_coords()).sum::<usize>()
+        self.exterior().coords_count() +
+            self.interiors().iter().map(|i| i.coords_count()).sum::<usize>()
     }
 }
 
@@ -158,7 +158,7 @@ impl<'a, T: CoordinateType + 'a> CoordsIter<'a, T> for MultiLineString<T> {
 
     /// Return the number of coordinates in the `MultiLineString`.
     fn coords_count(&'a self) -> usize {
-        self.0.iter().map(|line_string| line_string.num_coords()).sum()
+        self.0.iter().map(|line_string| line_string.coords_count()).sum()
     }
 }