+++ geo-types/CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* `geo_types::LineString::num_coords` has been deprecated in favor of `geo::algorithm::coords_iter::CoordsIter::coords_count`
+  * <https://github.com/georust/geo/pull/563>
+
 ## 0.6.2
 
 * Add `into_iter`, `iter` and `iter_mut` methods for `MultiPolygon`, `MultiPoint`, and `MultiLineString`